join error_java.sql.SQLException : 부적합한 열 유형: getInt not implemented for class oracle.jdbc.driver.

+++ src/main/java/hello/hellospring/repository/MemoryMemberRepository.java
@@ -7,8 +7,8 @@ import java.util.*;
 
 public class MemoryMemberRepository implements MemberRepository{
 
-    private static Map<Long, Member> store = new HashMap<>();
-    private static long sequence = 0L;
+    private static Map<Integer, Member> store = new HashMap<>();
+    private static int sequence = 0;
 
     @Override
     public Member save(Member member) {
@@ -18,7 +18,7 @@ public class MemoryMemberRepository implements MemberRepository{
     }
 
     @Override
-    public Optional<Member> findById(Long id) {
+    public Optional<Member> findById(int id) {
 
         return Optional.ofNullable(store.get(id));
     }
