+++ src/main/java/hello/hellospring/service/MemberService.java
@@ -22,7 +22,7 @@ public class MemberService {
     /**
      * 회원가입
      **/
-    public Long join(Member member) {
+    public Integer join(Member member) {
 
         validateDuplicateMember(member); //중복 회원 검증
         memberRepository.save(member);
@@ -43,7 +43,7 @@ public class MemberService {
         return memberRepository.findAll();
     }
 
-    public Optional<Member> findOne(Long memberId) {
+    public Optional<Member> findOne(int memberId) {
         return memberRepository.findById(memberId);
     }
 }
